+++ outputs-module.tf
@@ -1,0 +1,9 @@
+output "terraform_module" {
+  description = "Information about this Terraform module"
+  value = {
+    name       = "db-sql"
+    version    = file("${path.module}/VERSION")
+    provider   = "azurerm"
+    maintainer = "claranet"
+  }
+}